add the new createTables to the rest of the tests that call db
