+++ packages/component-dashboard/server/resolvers.test.js
@@ -1,5 +1,5 @@
 jest.mock('@pubsweet/logger')
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const mailer = require('@pubsweet/component-send-email')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
