+++ packages/component-model-audit-log/entities/auditLog/index.test.js
@@ -1,4 +1,4 @@
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const uuid = require('uuid')
 const AuditLog = require('.')
 const replaySetup = require('../../../../test/helpers/replay-setup')
