+++ packages/component-model-user/resolvers.test.js
@@ -1,4 +1,4 @@
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const replaySetup = require('../../test/helpers/replay-setup')
 const { Query } = require('./resolvers')
