+++ packages/component-model/index.js
@@ -1,1 +1,6 @@
 // DO NOT REMOVE (see README for more info).
+const createTables = require('./test-helpers/createTables')
+
+module.exports = {
+  createTables,
+}
