+++ packages/component-submission/server/resolvers/updateManuscript.test.js
@@ -1,6 +1,6 @@
 jest.mock('@pubsweet/logger')
 
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
 const { Mutation } = require('.')
