+++ packages/component-submission/server/resolvers/uploadManuscript.test.js
@@ -8,7 +8,7 @@ const config = require('config')
 const fs = require('fs-extra')
 const stream = require('stream')
 const logger = require('@pubsweet/logger')
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const mailer = require('@pubsweet/component-send-email')
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
