+++ packages/component-submission/server/resolvers/uploadSupportingFile.test.js
@@ -1,6 +1,6 @@
 const fs = require('fs-extra')
 const config = require('config')
-const { createTables } = require('@pubsweet/db-manager')
+const { createTables } = require('@elifesciences/component-model')
 const AuditLog = require('@elifesciences/component-model-audit-log').model
 const User = require('@elifesciences/component-model-user').model
 const Manuscript = require('@elifesciences/component-model-manuscript').model
